+++ src/main/java/com/barco/model/dto/request/ControlRequest.java
@@ -20,12 +20,10 @@ public class ControlRequest extends RequestFilter {
     private String fieldTitle;
     private String fieldName;
     private String placeHolder;
-    private Long fieldWidth;
     private Long minLength;
     private Long maxLength;
     private String fieldLkValue;
     private Long mandatory;
-    private Long disabled;
     private Long isDefault;
     private String defaultValue;
     private String pattern;
@@ -37,6 +35,7 @@ public class ControlRequest extends RequestFilter {
     private List<Long> controlLinkSection;
     private String disabledPattern;
     private String visiblePattern;
+    private Long fieldWidth;
     private Long controlOrder;
 
     public ControlRequest() {
@@ -98,14 +97,6 @@ public class ControlRequest extends RequestFilter {
         this.placeHolder = placeHolder;
     }
 
-    public Long getFieldWidth() {
-        return fieldWidth;
-    }
-
-    public void setFieldWidth(Long fieldWidth) {
-        this.fieldWidth = fieldWidth;
-    }
-
     public Long getMinLength() {
         return minLength;
     }
@@ -136,14 +127,6 @@ public class ControlRequest extends RequestFilter {
 
     public void setMandatory(Long mandatory) {
         this.mandatory = mandatory;
-    }
-
-    public Long getDisabled() {
-        return disabled;
-    }
-
-    public void setDisabled(Long disabled) {
-        this.disabled = disabled;
     }
 
     public Long getIsDefault() {
@@ -224,6 +207,14 @@ public class ControlRequest extends RequestFilter {
 
     public void setVisiblePattern(String visiblePattern) {
         this.visiblePattern = visiblePattern;
+    }
+
+    public Long getFieldWidth() {
+        return fieldWidth;
+    }
+
+    public void setFieldWidth(Long fieldWidth) {
+        this.fieldWidth = fieldWidth;
     }
 
     public Long getControlOrder() {
