+++ src/main/java/com/barco/model/dto/request/LookupDataRequest.java
@@ -26,6 +26,10 @@ public class LookupDataRequest {
     public LookupDataRequest() {
     }
 
+    public LookupDataRequest(String lookupType) {
+        this.lookupType = lookupType;
+    }
+
     public Long getId() {
         return id;
     }
